refactor(views): extract shared Attributes, ActionsButtons, and ErrorPage
DRY up duplicated update_attributes into a Views::Attributes module
shared by BasicLayout and BasicComponent. Extract inline actions_buttons
methods from Index and Show into a reusable ActionsButtons component.
Consolidate three identical error page view_templates into an ErrorPage
base class. Also adds nil guard on field access in Index and Show
table/record iteration.
