+++ lib/tiny_admin/views/attributes.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module TinyAdmin
+  module Views
+    module Attributes
+      def update_attributes(attributes)
+        attributes.each do |key, value|
+          send("#{key}=", value)
+        end
+      end
+    end
+  end
+end
