+++ lib/tiny_admin/views/basic_layout.rb
@@ -3,18 +3,13 @@
 module TinyAdmin
   module Views
     class BasicLayout < Phlex::HTML
+      include Attributes
       include Utils
 
       attr_accessor :content, :params, :widgets
 
       def label_for(value, options: [])
         TinyAdmin.settings.helper_class.label_for(value, options: options)
-      end
-
-      def update_attributes(attributes)
-        attributes.each do |key, value|
-          send("#{key}=", value)
-        end
       end
     end
   end
