+++ lib/tiny_admin/views/pages/page_not_found.rb
@@ -3,15 +3,7 @@
 module TinyAdmin
   module Views
     module Pages
-      class PageNotFound < DefaultLayout
-        def view_template
-          super do
-            div(class: "page_not_found") {
-              h1(class: "title") { title }
-            }
-          end
-        end
-
+      class PageNotFound < ErrorPage
         def title
           "Page not found"
         end
