+++ lib/tiny_admin/views/pages/record_not_found.rb
@@ -3,15 +3,7 @@
 module TinyAdmin
   module Views
     module Pages
-      class RecordNotFound < DefaultLayout
-        def view_template
-          super do
-            div(class: "record_not_found") {
-              h1(class: "title") { title }
-            }
-          end
-        end
-
+      class RecordNotFound < ErrorPage
         def title
           "Record not found"
         end
